backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
google-adk
google-genai
psycopg2-binary